refactor: add :unit:mentions module
