+++ feature/inbox/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/feature/inbox/di/InboxModule.kt
@@ -3,8 +3,6 @@ package com.github.diegoberaldin.raccoonforlemmy.feature.inbox.di
 import com.github.diegoberaldin.raccoonforlemmy.core.architecture.DefaultMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.main.InboxMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.main.InboxViewModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.mentions.InboxMentionsMviModel
-import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.mentions.InboxMentionsViewModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.messages.InboxMessagesMviModel
 import com.github.diegoberaldin.raccoonforlemmy.feature.inbox.messages.InboxMessagesViewModel
 import org.koin.dsl.module
@@ -17,19 +15,6 @@ val inboxTabModule = module {
             userRepository = get(),
             coordinator = get(),
             settingsRepository = get(),
-            notificationCenter = get(),
-        )
-    }
-    factory<InboxMentionsMviModel> {
-        InboxMentionsViewModel(
-            mvi = DefaultMviModel(InboxMentionsMviModel.UiState()),
-            userRepository = get(),
-            identityRepository = get(),
-            commentRepository = get(),
-            themeRepository = get(),
-            settingsRepository = get(),
-            hapticFeedback = get(),
-            coordinator = get(),
             notificationCenter = get(),
         )
     }
