+++ settings.gradle.kts
@@ -78,3 +78,4 @@ include(":unit:userdetail")
 include(":unit:web")
 include(":unit:zoomableimage")
 include(":unit:replies")
+include(":unit:mentions")
