+++ unit/mentions/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/mentions/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.mentions.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.mentions.InboxMentionsMviModel
+
+expect fun getInboxMentionsViewModel(): InboxMentionsMviModel
