+++ unit/mentions/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/mentions/di/Utils.kt
@@ -1,0 +1,11 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.mentions.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.mentions.InboxMentionsMviModel
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+
+actual fun getInboxMentionsViewModel(): InboxMentionsMviModel = UnitMentionsDiHelper.mentionsModel
+
+object UnitMentionsDiHelper : KoinComponent {
+    val mentionsModel: InboxMentionsMviModel by inject()
+}
